requirements.txt
requests>=2.31.0
sseclient-py>=1.8.0
python-dotenv>=1.0.0
pytest>=7.4.0
ruff>=0.4.0
